revert code reorganize, postpone to future pr
